feat: offload layout processing to worker
